move MarkdownTextArea outside pages; add proptypes for it
